+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2018-05-22
+
+### Changed
+
+- Use a third-party package to validate the configuration file YAML instead of custom code. Closes #66.
+
 ## [0.6.0] - 2018-05-21
 
 ### Added